+++ docs/src/release_notes.rst
@@ -7,10 +7,11 @@ This release notes format is based on
 `Keep a Changelog <https://keepachangelog.com>`_.
 Mold adheres to `Semantic Versioning <https://semver.org>`_.
 
-Unreleased
-----------
+0.2.0 (2021-09-12)
+------------------
 - Prefill configuration values
 - Add more licenses
+- Improve builtin configurations and plugins
 
 0.1.0 (2021-08-16)
 ------------------